Handle firestations down with cache approach on views, implement consensus between pluviometers

+++ src/main/scala/it/unibo/pcd/distributed/model/Pluviometer.scala
@@ -2,4 +2,4 @@ package it.unibo.pcd.distributed.model
 
 case class Pluviometer(zoneId: Int,
                        position: Point2D,
-                       threshold: Float)
+                       threshold: Double)
